requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
aiosqlite==0.19.0
alembic==1.11.3
asttokens==2.4.1
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.1.0
beautifulsoup4==4.12.2
black==23.12.1
braceexpand==0.1.7
cachetools==5.3.2
certifi==2023.11.17
click==8.1.7
comm==0.2.0
debugpy==1.8.0
decorator==5.1.1
discord.py==2.3.2
enkanetwork.py @ git+https://github.com/KT-Yeh/EnkaNetwork.py
exceptiongroup==1.2.0
executing==2.0.1
flake8==6.1.0
frozenlist==1.4.1
genshin @ git+https://github.com/KT-Yeh/genshin.py
greenlet==3.0.3
idna==3.6
import-expression==1.1.4
ipykernel==6.27.1
ipython==8.19.0
jedi==0.19.1
jishaku==2.5.2
jupyter_client==8.6.0
jupyter_core==5.5.1
Mako==1.3.0
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
mccabe==0.7.0
mihomo @ git+https://github.com/KT-Yeh/mihomo@v1.1.7
multidict==6.0.4
mypy==1.8.0
mypy-extensions==1.0.0
nest-asyncio==1.5.8
packaging==23.2
parso==0.8.3
pathspec==0.12.1
Pillow==9.5.0
platformdirs==4.1.0
prometheus-client==0.19.0
prompt-toolkit==3.0.43
psutil==5.9.7
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.11.1
pydantic==1.10.13
pyflakes==3.1.0
Pygments==2.17.2
python-dateutil==2.8.2
python-dotenv==1.0.0
pyzmq==25.1.2
sentry-sdk==1.39.1
six==1.16.0
soupsieve==2.5
SQLAlchemy==2.0.23
stack-data==0.6.3
tomli==2.0.1
tornado==6.4
traitlets==5.14.0
typing_extensions==4.9.0
urllib3==2.1.0
wcwidth==0.2.12
yarl==1.9.4